extract interface
